requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.5.7
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.18.26
botocore==1.21.26
CacheControl==0.12.10
cachetools==5.0.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
django-crontab==0.7.1
django-filter==21.1
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
docopt==0.6.2
firebase-admin==5.2.0
google-api-core==2.5.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.2
jmespath==0.10.0
kombu==5.1.0
msgpack==1.0.3
num2words==0.5.10
packaging==21.3
pdfkit==1.0.0
prompt-toolkit==3.0.20
proto-plus==1.20.3
protobuf==3.19.4
psycopg2==2.9.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycurl==7.43.0.5
PyJWT==2.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.1
razorpay==1.2.0
requests==2.26.0
rsa==4.8
s3transfer==0.5.0
sentry-sdk==1.4.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
